bts/Transceiver.hs: Make a start on the main loops of Layer0
The Transceiver is the 'SAP' of Layer0 where all of Layer1's
requests are handled via a pair of UDP sockets, one for data
and the other for control. Transceiver serves as the library
of actions that run as independant threads from Main.hs where
each thread services a particular subsystem of Layer0.

Signed-off-by: Edward O'Callaghan <eocallaghan@alterapraxis.com>